01-Introducao-ML/data/ieee-cis-fraud/processed/requirements.txt
# Requisitos para Processamento de Dados e IA
pandas>=2.0.0
numpy>=1.23.0
pyarrow>=12.0.0
scikit-learn>=1.2.0
openai==0.28.1  # Versão fixa para compatibilidade
httpx==0.23.0  # Versão fixa para evitar erro de SUPPORTED_DECODERS
tqdm>=4.65.0
python-dotenv>=1.0.0

# Ferramentas adicionais
jupyter>=1.0.0
ipykernel>=6.0.0
httpx==0.24.1
